Refactor autocomplete into own component
