+++ components/search/AutoSuggest.vue
@@ -1,119 +1,58 @@
 <template>
-  <b-form
-    ref="form"
-    inline
-    @submit.prevent="submitForm"
+  <b-list-group
+    v-show="isActive"
+    :id="elementId"
+    class="auto-suggest-dropdown"
+    data-qa="search suggestions"
+    :aria-hidden="!isActive"
   >
-    <b-input-group
-      role="combobox"
-      aria-owns="autocomplete-results"
-      :aria-expanded="isActive"
-      class="auto-suggest"
+    <b-list-group-item
+      v-if="isLoading"
+      class="loading"
+    >
+      {{ $t('loadingResults') }}{{ $t('formatting.ellipsis') }}
+    </b-list-group-item>
+
+    <b-list-group-item
+      v-for="(val, name, index) in value"
+      v-else
+      :key="index"
+      role="option"
+      :aria-selected="index === focus"
+      :to="linkGen(name)"
+      :class="{ 'hover': index === focus }"
+      :data-index="index"
+      @mouseover="focus = index"
+      @focus="index === focus"
+      @click="clickSuggestion"
     >
       <template
-        v-if="pillLabel"
-        v-slot:prepend
+        v-for="(part, partIndex) in highlightResult(val)"
       >
-        <SearchBarPill
-          :text="pillLabel"
-          :remove-link-label="$t('removeFilter', { filterLabel: pillLabel })"
-          :remove-link-to="removeLinkTo"
-        />
-      </template>
-      <b-form-input
-        ref="searchbox"
-        v-model="query"
-        :autocomplete="enableAutoSuggest ? 'off' : 'on'"
-        :placeholder="$t('searchPlaceholder')"
-        name="query"
-        data-qa="search box"
-        role="searchbox"
-        aria-autocomplete="list"
-        aria-controls="autocomplete-results"
-        :aria-label="$t('search')"
-        @input="searchboxInput"
-        @focus="activateDropdown"
-      />
-      <b-button
-        type="submit"
-        data-qa="search button"
-        variant="primary"
-      >
-        <span class="sr-only">
-          {{ $t('search') }}
-        </span>
-        <img
-          src="../../assets/img/magnifier.svg"
-          :alt="$t('search')"
-        >
-      </b-button>
-      <b-list-group
-        v-show="isActive"
-        id="autocomplete-results"
-        class="auto-suggest-dropdown"
-        data-qa="search suggestions"
-        :aria-hidden="!isActive"
-      >
-        <b-list-group-item
-          v-if="isLoading"
-          class="loading"
-        >
-          {{ $t('loadingResults') }}{{ $t('formatting.ellipsis') }}
-        </b-list-group-item>
-
-        <b-list-group-item
-          v-for="(value, name, index) in suggestions"
+        <strong
+          v-if="part.highlight"
+          :key="partIndex"
+          class="highlight"
+          data-qa="highlighted"
+        >{{ part.text }}</strong> <!-- Do not put onto a new line -->
+        <span
           v-else
-          :key="index"
-          role="option"
-          :aria-selected="index === focus"
-          :to="suggestionLinkGen(name)"
-          :class="{ 'hover': index === focus }"
-          :value="localiseSuggestionLabel(value)"
-          :data-index="index"
-          @mouseover="focus = index"
-          @focus="index === focus"
-          @click="closeDropdown"
-        >
-          <template
-            v-for="(part, partIndex) in highlightResult(value)"
-          >
-            <strong
-              v-if="part.highlight"
-              :key="partIndex"
-              class="highlight"
-              data-qa="highlighted"
-            >{{ part.text }}</strong> <!-- Do not put onto a new line -->
-            <span
-              v-else
-              :key="partIndex"
-              data-qa="base"
-            >{{ part.text }}</span> <!-- Do not put onto a new line -->
-          </template>
-        </b-list-group-item>
-      </b-list-group>
-    </b-input-group>
-  </b-form>
+          :key="partIndex"
+          data-qa="base"
+        >{{ part.text }}</span> <!-- Do not put onto a new line -->
+      </template>
+    </b-list-group-item>
+  </b-list-group>
 </template>
 
 <script>
-  import SearchBarPill from './SearchBarPill.vue';
   import match from 'autosuggest-highlight/match';
   import parse from 'autosuggest-highlight/parse';
 
   export default {
     name: 'AutoSuggest',
 
-    components: {
-      SearchBarPill
-    },
-
     props: {
-      enableAutoSuggest: {
-        type: Boolean,
-        default: false
-      },
-
       // Property names are identifiers, emitted when suggestion is selected.
       // Property values are lang maps for labels to display.
       // @example
@@ -127,20 +66,34 @@
       //         es: 'góspel'
       //       }
       //     }
-      suggestions: {
+      value: {
         type: Object,
         default: () => {}
       },
 
-      suggestionLinkGen: {
+      query: {
+        type: String,
+        default: ''
+      },
+
+      linkGen: {
         type: Function,
         default: (val) => val
+      },
+
+      elementId: {
+        type: String,
+        default: 'autocomplete-results'
+      },
+
+      inputRefName: {
+        type: String,
+        default: 'searchbox'
       }
     },
 
     data() {
       return {
-        query: null,
         focus: null,
         isActive: false,
         isLoading: false
@@ -152,55 +105,53 @@
         return this.$store.state.i18n.locale;
       },
 
-      onSearchablePage() {
-        return this.$store.state.search.active;
+      numberOfSuggestions() {
+        return Object.keys(this.value).length;
       },
 
-      pillLabel() {
-        return this.$store.state.search.pill;
+      noSuggestionHasFocus() {
+        return this.focus === null;
       },
 
-      routePath() {
-        if (this.onSearchablePage) {
-          return this.$route.path;
-        }
-        return this.localePath({ name: 'search' });
+      firstSuggestionHasFocus() {
+        return this.focus === 0;
       },
 
-      removeLinkTo() {
-        return {
-          path: this.localePath({
-            name: 'search'
-          }),
-          query: { ...this.$route.query, page: 1 }
-        };
-      },
-
-      view() {
-        return this.$store.getters['search/activeView'];
+      lastSuggestionHasFocus() {
+        return this.focus === (this.numberOfSuggestions - 1);
       }
     },
 
     watch: {
-      suggestions() {
-        this.isActive = !!this.suggestions;
-        this.isLoading = false;
-      },
       '$route.query'() {
-        this.queryOnSearchablePage();
         this.closeDropdown();
+      },
+
+      value() {
+        this.isActive = true;
       }
     },
 
     mounted() {
-      if (!this.enableAutoSuggest) return;
-
-      this.queryOnSearchablePage();
-      document.addEventListener('keyup', this.navigateDropdown);
+      this.$parent.$refs[this.inputRefName].$el.addEventListener('keyup', this.navigateDropdown);
       document.addEventListener('mouseup', this.clickOutside);
     },
 
     methods: {
+      clickOutside(event) {
+        if (!this.isActive) return;
+        const isChild = this.$el.contains(event.target);
+        if (!isChild) {
+          this.closeDropdown();
+        }
+      },
+
+      // Localise a lang map
+      //
+      // Order of priority:
+      // 1. User's UI language
+      // 2. English
+      // 3. First available value
       localiseSuggestionLabel(value) {
         if (value[this.locale]) {
           return value[this.locale];
@@ -210,11 +161,29 @@
         return Object.values(value)[0];
       },
 
+      // Highlight the user's query in a suggestion
       highlightResult(value) {
-        const string = this.localiseSuggestionLabel(value);
-        const matches = match(string, this.query);
-        const parts = parse(string, matches);
-        return parts;
+        let matchingValues = {};
+
+        // Find all the suggestion labels that match the query
+        for (const locale in value) {
+          const string = value[locale];
+          const matches = match(string, this.query);
+          if (matches.length > 0) {
+            matchingValues[locale] = parse(string, matches);
+          }
+        }
+
+        // If any suggestions match, return the localised one with higlight
+        if (Object.values(matchingValues).length > 0) {
+          return this.localiseSuggestionLabel(matchingValues);
+        }
+
+        // No matches, so return a localised suggestion without highlight
+        return [{
+          text: this.localiseSuggestionLabel(value),
+          highlight: false
+        }];
       },
 
       closeDropdown() {
@@ -222,78 +191,54 @@
         this.focus = null;
       },
 
-      clickOutside(event) {
-        if (!this.isActive) return;
-
-        const isChild = this.$el.contains(event.target);
-
-        if (!isChild) {
-          this.closeDropdown();
-        }
+      clickSuggestion() {
+        const selected = Object.keys(this.value)[this.focus];
+        this.$emit('selected', selected);
+        this.closeDropdown();
       },
 
       focusOnSuggestion() {
         if (!this.focus) return;
-
-        const selectedSuggestion = this.$el.querySelector(`[data-index="${this.focus}"]`);
-        selectedSuggestion.focus();
+        // FIXME: this is problematic because it causes the view port to scroll
+        //        in addition to suggestion highlighting
+        // const selectedSuggestion = this.$el.querySelector(`[data-index="${this.focus}"]`);
+        // selectedSuggestion.focus();
       },
 
       navigateDropdown(event) {
         if (!this.isActive) return;
 
+        const selectedSuggestion = this.$el.querySelector(`[data-index="${this.focus}"]`);
+
         switch (event.keyCode) {
         case 9: // Tab Key
-          this.clickOutside(event);
+          this.closeDropdown();
+          break;
+        case 13: // Enter Key
+          // FIXME: this fails to intercept the form submission
+          event.preventDefault();
+          selectedSuggestion.click();
           break;
         case 27: // Escape Key
           this.closeDropdown();
           break;
         case 38: // Up Key
-          if (this.focus === null) {
-            this.focus = 0;
-          } else if (this.focus > 0) {
+          if (this.noSuggestionHasFocus || this.firstSuggestionHasFocus) {
+            this.focus = this.numberOfSuggestions - 1;
+          } else {
             this.focus--;
-          } else if (this.focus === 0) {
-            this.focus = null;
-            this.$refs.searchbox.focus();
           }
           this.focusOnSuggestion();
           break;
         case 40: // Down key
-          if (this.focus === null) {
+          if (this.noSuggestionHasFocus || this.lastSuggestionHasFocus) {
             this.focus = 0;
-          } else if (this.focus < Object.keys(this.suggestions).length - 1) {
+          } else {
             this.focus++;
           }
           this.focusOnSuggestion();
           break;
         }
-      },
-
-      activateDropdown() {
-        return !this.isActive && this.suggestions;
-      },
-
-      queryOnSearchablePage() {
-        this.onSearchablePage ? this.query = this.$store.state.search.query : this.query = '';
-      },
-
-      async submitForm() {
-        const newRouteQuery = { ...this.$route.query, ...{ query: this.query, page: 1, view: this.view } };
-        const newRoute = { path: this.routePath, query: newRouteQuery };
-        this.closeDropdown;
-        await this.$router.push(newRoute);
-      },
-
-      async searchboxInput() {
-        if (!this.enableAutoSuggest) return;
-        // Uncomment this to show a loading indicator when suggestions are
-        // awaiting update.
-        // TODO: decide if we want it or not. Entity API responses are so fast,
-        //       it's not really necessary.
-        // this.isLoading = true;
-        this.$emit('input', this.query);
       }
     }
   };
@@ -333,32 +278,6 @@
       .loading {
         font-size: 0.75rem;
       }
-    }
-  }
-
-  .input-group {
-    width: 100%;
-
-    .input-group-prepend {
-      align-items: center;
-      background-color: $lightgrey;
-      padding-left: .75rem;
-      padding-right: .1rem;
-      border-radius: 0.375rem 0 0 0.375rem;
-    }
-  }
-
-  .form-control {
-    background-color: $lightgrey;
-    border-radius: $border-radius 0 0 $border-radius;
-    margin-right: 0;
-  }
-
-  .btn {
-    border-radius: 0 $border-radius $border-radius 0;
-
-    img {
-      display: flex;
     }
   }
 </style>
